make source module gen a fencil processor
